requirements.txt
Django>=1.6,<1.7
cairocffi==0.6
CairoSVG==1.0.13
docker-compose==1.2.0
django-casper
